separated range-selector
