+++ projects/ngx-charts-on-fhir/src/lib/fhir-chart/fhir-chart.component.ts
@@ -33,11 +33,6 @@ export class FhirChartComponent implements OnInit {
 
   constructor(private configService: FhirChartConfigurationService, public layerManager: DataLayerManagerService) {}
 
-  getChartHeight(): string {
-    const rangeSelectorHeight = '72px';
-    return `calc(${this.height} - ${rangeSelectorHeight})`;
-  }
-
   ngOnInit(): void {
     Chart.register(annotationPlugin, zoomPlugin, scaleStackDividerPlugin);
     Chart.register(MedicationScale);
