+++ projects/ngx-charts-on-fhir/src/lib/range-selector/range-selector.component.html
@@ -9,13 +9,13 @@
     >
     <mat-button-toggle id="resetzoom" value="All" (click)="resetZoomChart()">All</mat-button-toggle>
   </mat-button-toggle-group>
-  <mat-form-field>
+  <mat-form-field subscriptSizing="dynamic">
     <mat-label>Start date</mat-label>
     <input matInput [matDatepicker]="picker" (dateChange)="dateChange($event, 'min')" [value]="minDate" id="mindate" />
     <mat-datepicker-toggle matSuffix [for]="picker"></mat-datepicker-toggle>
     <mat-datepicker #picker></mat-datepicker>
   </mat-form-field>
-  <mat-form-field>
+  <mat-form-field subscriptSizing="dynamic">
     <mat-label>End date</mat-label>
     <input matInput [matDatepicker]="pickerDate" (dateChange)="dateChange($event, 'max')" [value]="maxDate" id="maxdate" />
     <mat-datepicker-toggle matSuffix [for]="pickerDate"></mat-datepicker-toggle>
